backend/requirements.txt
Flask
Flask-Cors
SQLAlchemy
psycopg2-binary
python-dotenv
python-dateutil
matplotlib
reportlab
Werkzeug
PyJWT